READY TO SUBMIT - Exceeds
I forgot to mention this in the notes when I submitted the first time. The reviewer missed it and rejected it.

For the conditional error required to get exceeds, I modified the email field so that if it is typed in and then deleted and left blank, the user will get an error that says "*Email field is blank." If they have characters in the form but they are not formatted correctly, it will read "*Please enter a valid email address."

Lines 224 and 228 in the script.js file is where the error messages are written.